pyproject.toml
[tool.poetry]
name = "vendor-management-system"
version = "0.1.0"
description = ""
authors = ["Aman Pandey <aman2001mi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"
psycopg2 = "^2.9.9"
pillow = "^10.1.0"
pre-commit = "^3.5.0"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
